Initialize common Thorchain plugin
